added Composer

+++ Cosyses/Ubuntu/functions/package/clean-packages.sh
@@ -1,3 +1,10 @@
 #!/bin/bash -e
 
-apt-get autoremove --purge -y
+obsoletePackages=( $(apt-get --dry-run autoremove | grep -Po '^Remv \K[^ ]+' | cat) )
+
+if [[ "${#obsoletePackages[@]}" -gt 0 ]]; then
+  echo "Removing obsolete packages"
+  apt-get autoremove --purge -y
+else
+  echo "No packages to clean"
+fi
